Move plugins to seprate packages
